[build] Fix arch detection (#214)
* [build] Fix arch detection

On ppc64le deb machines, when _not_ cross compiling, `uname -m` reports
`ppc64le`, which is typically right. However, the "Debian machine
triplet" that is necessary to build correctly uses `powerpc64le` for
whatever reason. The cross compile path knows this and is explicit about
that.

To alleviate that, use `dpkg-architecture` to discover the HOST _CPU_
type, which is what forms the machine triplet. The conditional check for
`dpkg-architecture` should be safe and reports the right values for
x86 and arm, too; falls back to `uname -m` anyways.

* resor to to cc -dumpmachine

seems more apt

Signed-off-by: Tobias Pape <tobias@netshed.de>

---------

Signed-off-by: Tobias Pape <tobias@netshed.de>